+++ app/src/main/java/com/example/teacherforboss/data/api/UserApi.kt
@@ -4,12 +4,16 @@ import com.example.teacherforboss.data.model.request.LoginRequest
 import com.example.teacherforboss.data.model.request.SocialLoginRequest
 import com.example.teacherforboss.data.model.response.LoginResponse
 import com.example.teacherforboss.data.model.response.socialLoginResponse
-import com.example.teacherforboss.presentation.ui.auth.signup.api.EmailCheckRequest
-import com.example.teacherforboss.presentation.ui.auth.signup.api.EmailCheckResponse
-import com.example.teacherforboss.presentation.ui.auth.signup.api.EmailRequest
-import com.example.teacherforboss.presentation.ui.auth.signup.api.EmailResponse
-import com.example.teacherforboss.presentation.ui.auth.signup.api.SignupRequest
-import com.example.teacherforboss.presentation.ui.auth.signup.api.SignupResponse
+import com.example.teacherforboss.data.model.request.signup.EmailCheckRequest
+import com.example.teacherforboss.data.model.response.signup.EmailCheckResponse
+import com.example.teacherforboss.data.model.request.signup.EmailRequest
+import com.example.teacherforboss.data.model.response.signup.EmailResponse
+import com.example.teacherforboss.data.model.request.signup.SignupRequest
+import com.example.teacherforboss.data.model.response.signup.SignupResponse
+import com.example.teacherforboss.data.model.request.signup.PhoneCheckRequest
+import com.example.teacherforboss.data.model.response.signup.PhoneCheckResponse
+import com.example.teacherforboss.data.model.request.signup.PhoneRequest
+import com.example.teacherforboss.data.model.response.signup.PhoneResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.Header
@@ -57,4 +61,16 @@ interface UserApi {
         @Body signupRequest: SignupRequest
     )
     : Response<SignupResponse>
+
+    @POST("auth/phone")
+    suspend fun phoneUser(
+        @Body phoneRequest: PhoneRequest
+    )
+    :Response<PhoneResponse>
+
+    @POST("auth/phone/check")
+    suspend fun phoneCheck(
+        @Body phoneCheckRequest: PhoneCheckRequest
+    )
+    :Response<PhoneCheckResponse>
 }
