+++ app/src/main/java/com/example/teacherforboss/data/model/request/signup/PhoneCheckRequest.kt
@@ -1,4 +1,4 @@
-package com.example.teacherforboss.signup.api
+package com.example.teacherforboss.data.model.request.signup
 
 import com.google.gson.annotations.SerializedName
