--- app/src/main/java/com/example/teacherforboss/login/UserRepository.kt
@@ -1,49 +1,0 @@
-package com.example.teacherforboss.login
-
-import com.example.teacherforboss.signup.api.EmailCheckRequest
-import com.example.teacherforboss.signup.api.EmailCheckResponse
-import com.example.teacherforboss.signup.api.EmailRequest
-import com.example.teacherforboss.signup.api.EmailResponse
-import com.example.teacherforboss.signup.api.PhoneCheckRequest
-import com.example.teacherforboss.signup.api.PhoneCheckResponse
-import com.example.teacherforboss.signup.api.PhoneRequest
-import com.example.teacherforboss.signup.api.PhoneResponse
-import com.example.teacherforboss.signup.api.SignupRequest
-import com.example.teacherforboss.signup.api.SignupResponse
-import retrofit2.Response
-
-class UserRepository {
-    suspend fun loginUser(loginRequest: LoginRequest):
-            Response<LoginResponse>?{
-        return UserApi.getApi()?.loginUser(loginRequest=loginRequest)
-    }
-    fun loginReissue(refreshToken:String):
-            Response<LoginResponse>?{
-        return UserApi.getApi()?.loginReissue(refreshToken ="Bearer ${refreshToken}" )
-    }
-
-    suspend fun emailUser(emailRequest: EmailRequest):
-            Response<EmailResponse>?{
-        return UserApi.getApi()?.emailUser(emailRequest = emailRequest)
-    }
-
-    suspend fun emailCheck(emailCheckRequest: EmailCheckRequest):
-            Response<EmailCheckResponse>?{
-        return UserApi.getApi()?.emailCheck(emailCheckRequest = emailCheckRequest)
-    }
-
-    suspend fun signupUser(signupRequest: SignupRequest):
-            Response<SignupResponse>?{
-        return UserApi.getApi()?.signupUser(signupRequest = signupRequest)
-    }
-
-    suspend fun phoneUser(phoneRequest: PhoneRequest):
-            Response<PhoneResponse>?{
-        return UserApi.getApi()?.phoneUser(phoneRequest = phoneRequest)
-    }
-
-    suspend fun phoneCheck(phoneCheckRequest: PhoneCheckRequest):
-            Response<PhoneCheckResponse>?{
-        return UserApi.getApi()?.phoneCheck(phoneCheckRequest = phoneCheckRequest)
-    }
-}
